+++ be/src/runtime/row-batch.cc
@@ -74,7 +74,7 @@ RowBatch::RowBatch(const RowDescriptor& row_desc, const TRowBatch& input_batch,
         &decompressor);
     DCHECK(status.ok()) << status.GetErrorMsg();
 
-    int uncompressed_size = input_batch.uncompressed_size;
+    int64_t uncompressed_size = input_batch.uncompressed_size;
     DCHECK_NE(uncompressed_size, -1) << "RowBatch decompression failed";
     uint8_t* data = tuple_data_pool_->Allocate(uncompressed_size);
     status = decompressor->ProcessBlock(true, compressed_size, compressed_data,
@@ -178,11 +178,11 @@ int RowBatch::Serialize(TRowBatch* output_batch) {
     // Try compressing tuple_data to compression_scratch_, swap if compressed data is
     // smaller
     scoped_ptr<Codec> compressor;
-    Status status =
-        Codec::CreateCompressor(NULL, false, THdfsCompression::LZ4, &compressor);
+    Status status = Codec::CreateCompressor(NULL, false, THdfsCompression::LZ4,
+                                            &compressor);
     DCHECK(status.ok()) << status.GetErrorMsg();
 
-    int compressed_size = compressor->MaxOutputLen(size);
+    int64_t compressed_size = compressor->MaxOutputLen(size);
     if (compression_scratch_.size() < compressed_size) {
       compression_scratch_.resize(compressed_size);
     }
